+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Current Master
 
 - Add support for calls to `eth.sign`.
+- Moved account exporting within transitioning subview on the account detail view.
+- Added buttons to the account export process.
 
 ## 1.7.0 2016-04-29